requirements.txt
av==14.4.0
detectron2==0.6
ffmpeg==1.4
ffmpeg_python==0.2.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
joblib==1.5.1
llava==0.0.1.dev0
matplotlib==3.10.5
numpy==1.24.4
openai==1.101.0
opencv_python==4.11.0.86
opencv_python_headless==4.5.5.64
pandas==2.3.2
Pillow==11.3.0
regex==2024.11.6
salesforce_lavis==1.0.2
scenedetect==0.6.6
scikit_learn==1.7.1
scipy==1.16.1
setuptools==78.1.1
simplejson==3.20.1
torch==1.13.1
torchtext==0.14.1
torchvision==0.14.1
tqdm==4.67.1
transformers==4.26.1
